+++ CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Template in request body
 - Attribute condition fhir path
 - Attribute join fhir path
+- Attribute condition id and value fhir path
 
 ### Changed
 - Rename Conversion Template as Converter Template